fix(header): refactor a bit more and remove old files
